requirements.txt
cohere==5.5.7
Flask==3.0.3
python-dotenv==0.19.2
